[frontend] Put headless frontend in its own directory.
